# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Humanizer AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface RegisterBody {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserPlan = (typeof UserPlan)[keyof typeof UserPlan];

export const UserPlan = {
  free: "free",
  basic: "basic",
  pro: "pro",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  isPremium: boolean;
  plan: UserPlan;
  usageCount: number;
  freeLimit: number;
  createdAt?: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface HumanizeBody {
  /**
   * @minLength 10
   * @maxLength 5000
   */
  text: string;
}

export interface HumanizeResponse {
  humanizedText: string;
  usageCount: number;
  freeLimit: number;
  isPremium: boolean;
}

export type CreateCheckoutBodyPlan =
  (typeof CreateCheckoutBodyPlan)[keyof typeof CreateCheckoutBodyPlan];

export const CreateCheckoutBodyPlan = {
  basic: "basic",
  pro: "pro",
} as const;

export interface CreateCheckoutBody {
  plan: CreateCheckoutBodyPlan;
}

export interface CheckoutResponse {
  url: string;
}