# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Humanizer AI API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register new user
 */
export const registerBodyPasswordMin = 6;

export const RegisterBody = zod.object({
  email: zod.string().email(),
  password: zod.string().min(registerBodyPasswordMin),
  name: zod.string(),
});

/**
 * @summary Login user
 */
export const LoginBody = zod.object({
  email: zod.string().email(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  token: zod.string(),
  user: zod.object({
    id: zod.number(),
    name: zod.string(),
    email: zod.string(),
    isPremium: zod.boolean(),
    plan: zod.enum(["free", "basic", "pro"]),
    usageCount: zod.number(),
    freeLimit: zod.number(),
    createdAt: zod.coerce.date().optional(),
  }),
});

/**
 * @summary Get current user profile
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  email: zod.string(),
  isPremium: zod.boolean(),
  plan: zod.enum(["free", "basic", "pro"]),
  usageCount: zod.number(),
  freeLimit: zod.number(),
  createdAt: zod.coerce.date().optional(),
});

/**
 * @summary Humanize AI-generated text
 */
export const humanizeTextBodyTextMin = 10;
export const humanizeTextBodyTextMax = 5000;

export const HumanizeTextBody = zod.object({
  text: zod.string().min(humanizeTextBodyTextMin).max(humanizeTextBodyTextMax),
});

export const HumanizeTextResponse = zod.object({
  humanizedText: zod.string(),
  usageCount: zod.number(),
  freeLimit: zod.number(),
  isPremium: zod.boolean(),
});

/**
 * @summary Create Stripe checkout session
 */
export const CreateCheckoutSessionBody = zod.object({
  plan: zod.enum(["basic", "pro"]),
});

export const CreateCheckoutSessionResponse = zod.object({
  url: zod.string(),
});

/**
 * @summary Stripe webhook handler
 */
export const StripeWebhookResponse = zod.object({
  success: zod.boolean(),
});