# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Humanizer AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPlan } from "./userPlan";

export interface User {
  id: number;
  name: string;
  email: string;
  isPremium: boolean;
  plan: UserPlan;
  usageCount: number;
  freeLimit: number;
  createdAt?: Date;
}